fix: adjustments (int -> Long) and table relations between entities

+++ src/main/java/com/fmt/educonnect/controllers/dtos/requests/RequestAlunoDTO.java
@@ -7,5 +7,5 @@ import java.time.LocalDate;
 public record RequestAlunoDTO(
         String nome,
         @JsonFormat(pattern = "dd-MM-yyyy") LocalDate dataNascimento,
-        Integer idCadastro) {
+        Long idCadastro) {
 }
