+++ src/main/java/com/fmt/educonnect/controllers/dtos/responses/ResponseDocenteDTO.java
@@ -3,10 +3,10 @@ package com.fmt.educonnect.controllers.dtos.responses;
 import java.time.LocalDate;
 
 public record ResponseDocenteDTO (
-        Integer id,
+        Long id,
         String nome,
         LocalDate dataEntrada,
-        int idCadastro
+        Long idCadastro
     ) {
 
 }
